requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic_settings
pydantic
python-jose
passlib[bcrypt]
requests